requirements.txt
streamlit
scrapegraphai
playwright
langchain-core
langchain-groq
langchain-openai
python-dotenv